# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JISHLink Workforce & Manpower Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  supabase?: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface EmployeeLoginRequest {
  employee_code: string;
  password: string;
}

export type SendOtpRequestPurpose = typeof SendOtpRequestPurpose[keyof typeof SendOtpRequestPurpose];


export const SendOtpRequestPurpose = {
  login: 'login',
  password_reset: 'password_reset',
} as const;

export interface SendOtpRequest {
  employee_code: string;
  purpose: SendOtpRequestPurpose;
}

export interface ForgotPasswordRequest {
  employee_code: string;
}

export interface ResetPasswordRequest {
  employee_code: string;
  reset_code: string;
  new_password: string;
}

export type VerifyOtpRequestPurpose = typeof VerifyOtpRequestPurpose[keyof typeof VerifyOtpRequestPurpose];


export const VerifyOtpRequestPurpose = {
  login: 'login',
  password_reset: 'password_reset',
} as const;

export interface VerifyOtpRequest {
  employee_code: string;
  otp: string;
  purpose: VerifyOtpRequestPurpose;
}

export interface OtpSentResponse {
  success: boolean;
  message?: string;
  expires_in_minutes?: number;
}

export type EmployeeProfileStatus = typeof EmployeeProfileStatus[keyof typeof EmployeeProfileStatus];


export const EmployeeProfileStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface EmployeeProfile {
  id: string;
  employee_code: string;
  name: string;
  mobile: string;
  email?: string | null;
  site_id?: string | null;
  client_id?: string | null;
  shift_id?: string | null;
  status: EmployeeProfileStatus;
  photo_url?: string | null;
}

export interface EmployeeLoginResponse {
  token: string;
  employee: EmployeeProfile;
}

export interface AttendanceScanRequest {
  site_token: string;
  latitude: number;
  longitude: number;
  device_id: string;
  timestamp: string;
}

export type AttendanceScanResponseAction = typeof AttendanceScanResponseAction[keyof typeof AttendanceScanResponseAction];


export const AttendanceScanResponseAction = {
  check_in: 'check_in',
  check_out: 'check_out',
} as const;

export type AttendanceScanResponseStatus = typeof AttendanceScanResponseStatus[keyof typeof AttendanceScanResponseStatus];


export const AttendanceScanResponseStatus = {
  present: 'present',
  late: 'late',
  early_out: 'early_out',
} as const;

export interface AttendanceScanResponse {
  success: boolean;
  action: AttendanceScanResponseAction;
  status: AttendanceScanResponseStatus;
  site_name: string;
  time: string;
  attendance_id?: string;
  distance_meters?: number;
}

export type AttendanceScanErrorCode = typeof AttendanceScanErrorCode[keyof typeof AttendanceScanErrorCode];


export const AttendanceScanErrorCode = {
  INVALID_QR: 'INVALID_QR',
  NOT_ASSIGNED: 'NOT_ASSIGNED',
  OUTSIDE_GEOFENCE: 'OUTSIDE_GEOFENCE',
  OUTSIDE_SHIFT_WINDOW: 'OUTSIDE_SHIFT_WINDOW',
  ALREADY_CHECKED_OUT: 'ALREADY_CHECKED_OUT',
} as const;

export interface AttendanceScanError {
  error: string;
  code: AttendanceScanErrorCode;
  distance_meters?: number;
  shift_start?: string;
  shift_end?: string;
}

export interface QrGenerateResponse {
  qr_code_url: string;
  qr_token: string;
  site_id: string;
  expires_at?: string | null;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  present: 'present',
  absent: 'absent',
  late: 'late',
  early_out: 'early_out',
} as const;

export interface AttendanceRecord {
  id: string;
  employee_id: string;
  employee_code: string;
  employee_name: string;
  site_id: string;
  site_name: string;
  date: string;
  check_in_time?: string | null;
  check_out_time?: string | null;
  status: AttendanceRecordStatus;
  check_in_lat?: number | null;
  check_in_lng?: number | null;
}

export interface DailyReportResponse {
  date: string;
  total: number;
  present: number;
  absent: number;
  late: number;
  early_out: number;
  records: AttendanceRecord[];
}

export interface MonthlyEmployeeRecord {
  employee_id: string;
  employee_code: string;
  employee_name: string;
  present_days: number;
  absent_days: number;
  late_days: number;
  attendance_percentage: number;
}

export interface MonthlyReportResponse {
  year: number;
  month: number;
  working_days: number;
  records: MonthlyEmployeeRecord[];
}

export interface DashboardStats {
  total_employees: number;
  present_today: number;
  absent_today: number;
  late_today: number;
  total_sites: number;
  total_clients: number;
}

export interface AttendanceDayStats {
  date: string;
  present: number;
  absent: number;
  late: number;
}

export interface AttendanceTrendResponse {
  data: AttendanceDayStats[];
}

export type RecentCheckinStatus = typeof RecentCheckinStatus[keyof typeof RecentCheckinStatus];


export const RecentCheckinStatus = {
  present: 'present',
  late: 'late',
} as const;

export interface RecentCheckin {
  id: string;
  employee_name: string;
  employee_code: string;
  site_name: string;
  check_in_time: string;
  status: RecentCheckinStatus;
  photo_url?: string | null;
}

export interface RecentCheckinsResponse {
  data: RecentCheckin[];
}

export type GetDailyReportParams = {
date: string;
siteId?: string;
employeeId?: string;
recruiterId?: string;
};

export type GetMonthlyReportParams = {
year: number;
month: number;
siteId?: string;
employeeId?: string;
recruiterId?: string;
};

export type ExportExcelReportParams = {
type: ExportExcelReportType;
date?: string;
year?: number;
month?: number;
siteId?: string;
employeeId?: string;
recruiterId?: string;
};

export type ExportExcelReportType = typeof ExportExcelReportType[keyof typeof ExportExcelReportType];


export const ExportExcelReportType = {
  daily: 'daily',
  monthly: 'monthly',
} as const;

export type ExportPdfReportParams = {
type: ExportPdfReportType;
date?: string;
year?: number;
month?: number;
siteId?: string;
employeeId?: string;
recruiterId?: string;
};

export type ExportPdfReportType = typeof ExportPdfReportType[keyof typeof ExportPdfReportType];


export const ExportPdfReportType = {
  daily: 'daily',
  monthly: 'monthly',
} as const;

export type GetAttendanceTrendParams = {
siteId?: string;
};

export type GetRecentCheckinsParams = {
limit?: number;
};