# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JISHLink Workforce & Manpower Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/scan:
    post:
      operationId: attendanceScan
      tags: [attendance]
      summary: Process QR attendance scan — validates site token, geofence, shift window, and saves record
      security:
        - EmployeeBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceScanRequest"
      responses:
        "200":
          description: Attendance recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceScanResponse"
        "400":
          description: Validation failed (geofence, shift window, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceScanError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/attendance-trend:
    get:
      operationId: getAttendanceTrend
      tags: [dashboard]
      summary: Get last 14 days attendance trend for chart
      security:
        - AdminBearerAuth: []
      parameters:
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Attendance trend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceTrendResponse"
  /dashboard/recent-checkins:
    get:
      operationId: getRecentCheckins
      tags: [dashboard]
      summary: Get recent check-in events (polled every 15s for live feed)
      security:
        - AdminBearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent check-in records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentCheckinsResponse"
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
    EmployeeBearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    AttendanceScanRequest:
      type: object
      properties:
        site_token:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        device_id:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - site_token
        - latitude
        - longitude
        - device_id
        - timestamp
    AttendanceScanResponse:
      type: object
      properties:
        success:
          type: boolean
        action:
          type: string
          enum: [check_in, check_out]
        status:
          type: string
          enum: [present, late, early_out]
        site_name:
          type: string
        time:
          type: string
          format: date-time
        attendance_id:
          type: string
          format: uuid
        distance_meters:
          type: number
      required:
        - success
        - action
        - status
        - site_name
        - time
    AttendanceScanError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            - INVALID_QR
            - NOT_ASSIGNED
            - OUTSIDE_GEOFENCE
            - OUTSIDE_SHIFT_WINDOW
            - ALREADY_CHECKED_OUT
        distance_meters:
          type: number
        shift_start:
          type: string
        shift_end:
          type: string
      required:
        - error
        - code
...
    DashboardStats:
      type: object
      properties:
        total_employees:
          type: integer
        present_today:
          type: integer
        absent_today:
          type: integer
        late_today:
          type: integer
        total_sites:
          type: integer
        total_clients:
          type: integer
      required:
        - total_employees
        - present_today
        - absent_today
        - late_today
        - total_sites
        - total_clients
    AttendanceTrendResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceDayStats"
      required:
        - data
    AttendanceDayStats:
      type: object
      properties:
        date:
          type: string
          format: date
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
      required:
        - date
        - present
        - absent
        - late
    RecentCheckinsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RecentCheckin"
      required:
        - data
    RecentCheckin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_name:
          type: string
        employee_code:
          type: string
        site_name:
          type: string
        check_in_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [present, late]
        photo_url:
          type: string
          nullable: true
      required:
        - id
        - employee_name
        - employee_code
        - site_name
        - check_in_time
        - status